RavenDB-19234 moved location
