+++ src/Raven.Studio/typescript/components/pages/resources/cluster/clusterDebug/ClusterDebugSummary.scss
